Improved error reporting on login screen, ability to share logs (#49)
This PR reworks the `ErrorController` to be more useful and allow
providing detailed error information that the user can get to by
tapping.

It also adds the ability to export and share logs from the app via share
sheet. This is accessible both from the settings screen and the login
screen, and should help debug some login issues people have.