fix: DateOnlyCoder uses regex to strip time + timezone (#287)
For DateOnlyCoding, we assume the device local timezone when parsing.
For this, we now strip out everything except for the date from the input
string, and only then run the formatter. This produces a naive Date
object in the local timezone. When we send it back to the backend, we
only format the date component, so roundtripping should be safe in all
cases. 🤞

See https://github.com/paulgessinger/swift-paperless/issues/286
# Conflicts:
#	current_changelog.txt